requirements.txt
yfinance
praw
dotenv
pandas_ta
numpy==1.22.4
ntscraper
openai
google-genai